# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SaaS MVP API",
    "version": "0.1.0",
    "description": "API mínima para el SaaS MVP"
  },
  "servers": [
    { "url": "http://localhost:4000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" } }, "required": ["email","password"] }
            }
          }
        },
        "responses": { "200": { "description": "Token" } }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" }, "tenantId": { "type": "string" } }, "required": ["email","password","tenantId"] }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get current user",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "User object" } }
      }
    },
    "/projects": {
      "get": {
        "summary": "List projects for tenant",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Array of projects" } }
      },
      "post": {
        "summary": "Create project",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" }, "tenantId": { "type": "string" } }, "required": ["name"] }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    }
  }
}